+++ modules/dnn/src/tensorflow/tf_io.cpp
@@ -25,8 +25,8 @@ Implementation of various functions which are related to Tensorflow models readi
 
 #include "tf_io.hpp"
 
-#include "../caffe/caffe_io.hpp"
-#include "../caffe/glog_emulator.hpp"
+#include "../protobuf_io.hpp"
+#include "../glog_emulator.hpp"
 
 namespace cv {
 namespace dnn {
